airflow/dags/Scripts/Pipeline_Scripts/Grobid/requirements.txt
boto3
python-dotenv
# lxml
# xml
requests